Move commands to commands module
